A Linked List in Java Main file
This program was to illustrate the workings of a linked list in the java language. Also, noting that Java is a garbage collecting language, unlike C++. This program creates a linked list, and uses a while loop to insert items into the linked list. The items/data values increment by 12.